requirements.txt
cssselect2==0.4.1
lxml==4.6.3
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
pyhere==1.0.0
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.2
scipy==1.7.1
six==1.16.0
svglib==1.1.0
svgwrite==1.4.1
tinycss2==1.1.0
tsp-solver2==0.4.1
webencodings==0.5.1
